move things to level_2.go and include mental notes
